Refactor classes to go under state package
