--- chatkit-core/src/main/kotlin/com/pusher/chatkit/internal/State.kt
@@ -1,7 +1,0 @@
-package com.pusher.chatkit.internal
-
-import com.pusher.chatkit.rooms.internal.JoinedRoomsState
-
-internal data class State(
-    val joinedRoomsState: JoinedRoomsState
-)
