+++ chatkit-core/src/main/kotlin/com/pusher/chatkit/rooms/state/JoinedRoomsState.kt
@@ -1,4 +1,4 @@
-package com.pusher.chatkit.rooms.internal
+package com.pusher.chatkit.rooms.state
 
 internal data class JoinedRoomsState(
     val rooms: Map<String, JoinedRoomInternalType>,
