+++ chatkit-core/src/main/kotlin/com/pusher/chatkit/state/ChatkitState.kt
@@ -1,0 +1,15 @@
+package com.pusher.chatkit.state
+
+import com.pusher.chatkit.rooms.state.JoinedRoomsState
+
+internal data class ChatkitState(
+    val joinedRoomsState: JoinedRoomsState,
+    val auxiliaryState: AuxiliaryState
+)
+
+internal data class AuxiliaryState(
+    val version: Int,
+    val joinedRoomsReceived: ReducerLastChange<JoinedRoomsState>
+)
+
+internal data class ReducerLastChange<T>(val version: Int, val data: T)
